added chunk fusion option

+++ src/main.cpp
@@ -60,17 +60,17 @@ void write_image_chunked(const std::string &filename, int width, int height, int
     png_destroy_write_struct(&png, &info);
 }
 
-void chunk_limited(int width, int height, int x_min, int x_max, int y_min, int y_max, int max_iter, int chunk_size, int num_workers, int chunk_start, int chunk_end, std::string out_path, bool silent)
+void chunk_limited(int width, int height, int x_min, int x_max, int y_min, int y_max, int max_iter, int chunk_size, int num_workers, int chunk_start, int chunk_end, std::string chunk_path, bool silent)
 {
     std::cout << "Berechne nur Chunks von " << chunk_start << " bis " << chunk_end << std::endl;
-    std::cout << "Speichere Chunks in: " << out_path << std::endl;
+    std::cout << "Speichere Chunks in: " << chunk_path << std::endl;
 
-    fs::create_directory(out_path);
+    fs::create_directory(chunk_path);
     std::cout << "Generiere Mandelbrot-Menge..." << std::endl;
-    generate_mandelbrot_limited(width, height, x_min, x_max, y_min, y_max, max_iter, chunk_size, num_workers, chunk_start, chunk_end, out_path, silent);
+    generate_mandelbrot_limited(width, height, x_min, x_max, y_min, y_max, max_iter, chunk_size, num_workers, chunk_start, chunk_end, chunk_path, silent);
 }
 
-void chunk_unlimited(int width, int height, int x_min, int x_max, int y_min, int y_max, int max_iter, int chunk_size, int num_workers, std::string temp_dir, std::string filename, bool silent)
+void chunk_unlimited(int width, int height, int x_min, int x_max, int y_min, int y_max, int max_iter, int chunk_size, int num_workers, std::string temp_dir, std::string filename, bool silent, bool delete_cache)
 {
     std::cout << "Dateiname: " << filename << std::endl;
 
@@ -83,17 +83,21 @@ void chunk_unlimited(int width, int height, int x_min, int x_max, int y_min, int
     std::cout << "Füge Chunks zusammen und speichere Bild..." << std::endl;
     write_image_chunked(filename, width, height, chunk_size, temp_dir);
 
-    fs::remove_all(temp_dir);
+    if (delete_cache)
+    {
+        std::cout << "Lösche temporäre Chunks..." << std::endl;
+        fs::remove_all(temp_dir);
+    }
 }
 
-void chunk_intervall(int width, int height, int x_min, int x_max, int y_min, int y_max, int max_iter, int chunk_size, int num_workers, int intervall, std::string out_path, bool silent, int offset)
+void chunk_intervall(int width, int height, int x_min, int x_max, int y_min, int y_max, int max_iter, int chunk_size, int num_workers, int intervall, std::string chunk_path, bool silent, int offset)
 {
     std::cout << "Berechne Chunks in Intervallen von " << intervall << std::endl;
-    std::cout << "Speichere Chunks in: " << out_path << std::endl;
+    std::cout << "Speichere Chunks in: " << chunk_path << std::endl;
 
-    fs::create_directory(out_path);
+    fs::create_directory(chunk_path);
     std::cout << "Generiere Mandelbrot-Menge..." << std::endl;
-    generate_mandelbrot_intervall(width, height, x_min, x_max, y_min, y_max, max_iter, chunk_size, num_workers, intervall, out_path, silent, offset);
+    generate_mandelbrot_intervall(width, height, x_min, x_max, y_min, y_max, max_iter, chunk_size, num_workers, intervall, chunk_path, silent, offset);
 }
 
 int main(int argc, char **argv)
@@ -109,10 +113,12 @@ int main(int argc, char **argv)
     std::string filename = "mandelbrot.png";
     int chunk_start = -1;
     int chunk_end = -1;
-    std::string out_path = "chunks";
+    std::string chunk_path = "chunks";
     bool silent = false;
     int intervall = -1;
     int offset = 0;
+    bool fusion = false;
+    bool delete_cache = false;
 
     // Überprüfen, ob Argumente übergeben wurden
     // Kleine Helferfunktionen, um Duplikate zu vermeiden
@@ -157,7 +163,7 @@ int main(int argc, char **argv)
                 max_iter = nextIntArg(i);
             else if (arg == "--chunk_size")
                 chunk_size = nextIntArg(i);
-            else if (arg == "--num_workers")
+            else if ((arg == "--workers") || (arg == "-j"))
                 num_workers = nextIntArg(i);
             else if ((arg == "--filename"))
             {
@@ -172,17 +178,24 @@ int main(int argc, char **argv)
                 chunk_start = nextIntArg(i);
             else if (arg == "--chunk_end")
                 chunk_end = nextIntArg(i);
-            else if ((arg == "--out_path" || arg == "-o"))
+            else if ((arg == "--chunk_path" || arg == "-o"))
             {
                 if (++i >= argc)
                 {
-                    std::cerr << "Fehler: fehlender Wert für --out_path" << std::endl;
+                    std::cerr << "Fehler: fehlender Wert für --chunk_path" << std::endl;
                     std::exit(1);
                 }
-                out_path = argv[i];
+                chunk_path = argv[i];
             }
             else if ((arg == "--silent" || arg == "-s"))
                 silent = true;
+            else if (arg == "--fusion")
+                fusion = true;
+            else if ((arg == "--delete") || (arg == "-t"))
+            {
+                delete_cache = true;
+                std::cout << "Lösche temporäre Chunks nach dem Zusammenfügen..." << std::endl;
+            }
             else if (arg == "--intervall")
                 intervall = nextIntArg(i);
             else if (arg == "--offset")
@@ -200,13 +213,15 @@ int main(int argc, char **argv)
                           << "  --y_max N          Ende des y-Bereichs (Standard: 1.5)\n"
                           << "  --max_iter N       Iterationen (Standard: 100)\n"
                           << "  --chunk_size N     Chunk-Größe (Standard: 100)\n"
-                          << "  --num_workers N    Worker (Standard: 3)\n"
+                          << "  --workers N, -j N  Worker (Standard: 3)\n"
                           << "  --filename STR     Ausgabedatei (Standard: mandelbrot.png)\n"
                           << "  --intervall N      Chunk-Intervall (Standard: aus)\n"
                           << "  --offset N         Chunk-Offset (Standard: 0)\n"
                           << "  --chunk_start N    Startindex (Standard: aus)\n"
                           << "  --chunk_end N      Endindex (Standard: aus)\n"
-                          << "  --out_path, -o STR Speicherpfad für Chunks (Standard: chunks)\n";
+                          << "  --fusion           Füge Chunks zusammen und speichere Bild\n"
+                          << "  --delete, -t       Lösche temporäre Chunks nach dem Zusammenfügen\n"
+                          << "  --chunk_path, -o STR Speicherpfad für Chunks (Standard: chunks)\n";
                 return 0;
             }
             else
@@ -219,18 +234,22 @@ int main(int argc, char **argv)
 
     auto start_time = std::chrono::high_resolution_clock::now();
 
-    std::cout << "Bildgröße: " << width << "x" << height << std::endl;
-    std::cout << "Mandelbrot-Bereich: [" << x_min << ", " << x_max << "], [" << y_min << ", " << y_max << "]" << std::endl;
-    std::cout << "Maximale Iterationen: " << max_iter << std::endl;
-    std::cout << "Chunk-Größe: " << chunk_size << std::endl;
-    std::cout << "Anzahl der Worker: " << num_workers << std::endl;
-
     if (chunk_start != -1 && chunk_end != -1)
     {
-        chunk_limited(width, height, x_min, x_max, y_min, y_max, max_iter, chunk_size, num_workers, chunk_start, chunk_end, out_path, silent);
+        std::cout << "Bildgröße: " << width << "x" << height << std::endl;
+        std::cout << "Mandelbrot-Bereich: [" << x_min << ", " << x_max << "], [" << y_min << ", " << y_max << "]" << std::endl;
+        std::cout << "Maximale Iterationen: " << max_iter << std::endl;
+        std::cout << "Chunk-Größe: " << chunk_size << std::endl;
+        std::cout << "Anzahl der Worker: " << num_workers << std::endl;
+        chunk_limited(width, height, x_min, x_max, y_min, y_max, max_iter, chunk_size, num_workers, chunk_start, chunk_end, chunk_path, silent);
     }
     else if (chunk_start != -1 || chunk_end != -1)
     {
+        std::cout << "Bildgröße: " << width << "x" << height << std::endl;
+        std::cout << "Mandelbrot-Bereich: [" << x_min << ", " << x_max << "], [" << y_min << ", " << y_max << "]" << std::endl;
+        std::cout << "Maximale Iterationen: " << max_iter << std::endl;
+        std::cout << "Chunk-Größe: " << chunk_size << std::endl;
+        std::cout << "Anzahl der Worker: " << num_workers << std::endl;
         std::cerr << "Fehler: chunk_start und chunk_end müssen beide gesetzt sein." << std::endl;
         return 1;
     }
@@ -238,6 +257,11 @@ int main(int argc, char **argv)
     {
         if (intervall <= 0)
         {
+            std::cout << "Bildgröße: " << width << "x" << height << std::endl;
+            std::cout << "Mandelbrot-Bereich: [" << x_min << ", " << x_max << "], [" << y_min << ", " << y_max << "]" << std::endl;
+            std::cout << "Maximale Iterationen: " << max_iter << std::endl;
+            std::cout << "Chunk-Größe: " << chunk_size << std::endl;
+            std::cout << "Anzahl der Worker: " << num_workers << std::endl;
             std::cerr << "Fehler: Das Intervall muss größer als 0 sein." << std::endl;
             return 1;
         }
@@ -246,12 +270,24 @@ int main(int argc, char **argv)
             std::cout << "Intervall: " << intervall << std::endl;
             std::cout << "Offset: " << offset << std::endl;
 
-            chunk_intervall(width, height, x_min, x_max, y_min, y_max, max_iter, chunk_size, num_workers, intervall, out_path, silent, offset);
+            chunk_intervall(width, height, x_min, x_max, y_min, y_max, max_iter, chunk_size, num_workers, intervall, chunk_path, silent, offset);
         }
+    }
+    else if (fusion == true)
+    {
+        std::cout << "Bildgröße: " << width << "x" << height << std::endl;
+        std::cout << "Chunk-Größe: " << chunk_size << std::endl;
+        std::cout << "Füge Chunks zusammen und speichere Bild..." << std::endl;
+        write_image_chunked(filename, width, height, chunk_size, chunk_path);
     }
     else
     {
-        chunk_unlimited(width, height, x_min, x_max, y_min, y_max, max_iter, chunk_size, num_workers, "temp_chunks", filename, silent);
+        std::cout << "Bildgröße: " << width << "x" << height << std::endl;
+        std::cout << "Mandelbrot-Bereich: [" << x_min << ", " << x_max << "], [" << y_min << ", " << y_max << "]" << std::endl;
+        std::cout << "Maximale Iterationen: " << max_iter << std::endl;
+        std::cout << "Chunk-Größe: " << chunk_size << std::endl;
+        std::cout << "Anzahl der Worker: " << num_workers << std::endl;
+        chunk_unlimited(width, height, x_min, x_max, y_min, y_max, max_iter, chunk_size, num_workers, chunk_path, filename, silent, delete_cache);
     }
 
     auto end_time = std::chrono::high_resolution_clock::now();
